list2DF {backports}	R Documentation

Backport of list2DF for R < 4.0.0

Description

See the original description in base::list2DF. Note that the behaviour of list2DF() changed in R 4.2.0. Instead of recycling vectors to common length, an exception is raised. This backport implements the new behaviour, throwing an error message.

Usage

list2DF(x = list(), nrow = 0L)

Examples

# get function from namespace instead of possibly getting
# implementation shipped with recent R versions:
bp_list2DF = getFromNamespace("list2DF", "backports")

bp_list2DF(list(x = 1:3, y = 2:4))

[Package backports version 1.4.1 Index]