backports-package {backports}	R Documentation

backports: Reimplementations of Functions Introduced Since R-3.0.0

Description

Functions introduced or changed since R v3.0.0 are re-implemented in this package. The backports are conditionally exported in order to let R resolve the function name to either the implemented backport, or the respective base version, if available. Package developers can make use of new functions or arguments by selectively importing specific backports to support older installations.

Author(s)

Maintainer: Michel Lang michellang@gmail.com (ORCID)

Authors:

R Core Team

See Also

Useful links:

https://github.com/r-lib/backports

Report bugs at https://github.com/r-lib/backports/issues

[Package backports version 1.4.1 Index]